PARQUET-212: Add DirectWriterTest base class.
This adds convenience methods for writing to files using the
RecordConsumer API directly. This is useful for mimicing files from
other writers for compatibility tests.
